Move existing terraforming command
